refactor: split ConnectorCredentialsSection into focused components, centralize authTypeLabel
- Extract OAuthCredentialRow into its own file (was inline in 364-line file)
- Extract CredentialRow into its own file
- Move authTypeLabel to shared lib/providerLabels.ts utility
- Hoist hasMixedAuth computation outside map callback (avoid per-item recompute)
- ConnectorCredentialsSection.tsx now ~105 lines (was 364)
