+++ frontend/src/pages/agents/connectors/CredentialRow.tsx
@@ -1,0 +1,130 @@
+import { useState } from "react";
+import {
+  CheckCircle2,
+  Circle,
+  ExternalLink,
+  Plus,
+  Trash2,
+} from "lucide-react";
+import { Button } from "@/components/ui/button";
+import { Badge } from "@/components/ui/badge";
+import type { CredentialSummary } from "@/hooks/useCredentials";
+import { serviceDisplayName, authTypeLabel } from "@/lib/providerLabels";
+import type { RequiredCredential } from "@/hooks/useConnectorDetail";
+import { AddCredentialDialog } from "./AddCredentialDialog";
+import { RemoveCredentialDialog } from "./RemoveCredentialDialog";
+
+export function CredentialRow({
+  requiredCredential,
+  storedCredentials,
+}: {
+  requiredCredential: RequiredCredential;
+  storedCredentials: CredentialSummary[];
+}) {
+  const [addDialogOpen, setAddDialogOpen] = useState(false);
+  const [removeTarget, setRemoveTarget] = useState<CredentialSummary | null>(
+    null,
+  );
+
+  const isConnected = storedCredentials.length > 0;
+
+  return (
+    <>
+      <div className="rounded-lg border p-3">
+        <div className="flex items-center justify-between">
+          <div className="flex items-center gap-3">
+            {isConnected ? (
+              <CheckCircle2 className="size-5 shrink-0 text-green-600 dark:text-green-400" />
+            ) : (
+              <Circle className="text-muted-foreground size-5 shrink-0" />
+            )}
+            <div>
+              <div className="flex items-center gap-2">
+                <p className="text-sm font-medium">{serviceDisplayName(requiredCredential.service)}</p>
+                <Badge variant="outline" className="text-xs">
+                  {authTypeLabel(requiredCredential.auth_type)}
+                </Badge>
+              </div>
+              {requiredCredential.instructions_url && (
+                <a
+                  href={requiredCredential.instructions_url}
+                  target="_blank"
+                  rel="noopener noreferrer"
+                  className="text-muted-foreground hover:text-foreground mt-0.5 inline-flex items-center gap-1 text-xs"
+                >
+                  <ExternalLink className="size-3" />
+                  How to get this credential
+                </a>
+              )}
+            </div>
+          </div>
+          <div className="flex items-center gap-2">
+            <span
+              className={`text-xs font-medium ${
+                isConnected
+                  ? "text-green-600 dark:text-green-400"
+                  : "text-muted-foreground"
+              }`}
+            >
+              {isConnected ? "Connected" : "Not configured"}
+            </span>
+            <Button
+              variant="outline"
+              size="sm"
+              onClick={() => setAddDialogOpen(true)}
+            >
+              <Plus className="size-3" />
+              {isConnected ? "Add Another" : "Connect"}
+            </Button>
+          </div>
+        </div>
+
+        {storedCredentials.length > 0 && (
+          <div className="mt-3 space-y-2 border-t pt-3">
+            {storedCredentials.map((cred) => (
+              <div
+                key={cred.id}
+                className="bg-muted/50 flex items-center justify-between rounded-md px-3 py-2"
+              >
+                <div className="min-w-0">
+                  <p className="truncate text-sm">
+                    {cred.label ?? cred.service}
+                  </p>
+                  <p className="text-muted-foreground text-xs">
+                    Added{" "}
+                    {new Date(cred.created_at).toLocaleDateString()}
+                  </p>
+                </div>
+                <Button
+                  variant="ghost"
+                  size="sm"
+                  className="text-destructive hover:text-destructive"
+                  onClick={() => setRemoveTarget(cred)}
+                  aria-label={`Remove credential ${cred.label ?? cred.service}`}
+                >
+                  <Trash2 className="size-4" />
+                </Button>
+              </div>
+            ))}
+          </div>
+        )}
+      </div>
+
+      <AddCredentialDialog
+        open={addDialogOpen}
+        onOpenChange={setAddDialogOpen}
+        credential={requiredCredential}
+      />
+
+      {removeTarget && (
+        <RemoveCredentialDialog
+          open={!!removeTarget}
+          onOpenChange={(open) => {
+            if (!open) setRemoveTarget(null);
+          }}
+          credential={removeTarget}
+        />
+      )}
+    </>
+  );
+}
